requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid==2.3.3
autopep8==1.5.1
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Dance==3.0.0
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.7.0
Pillow==7.1.1
pluggy==0.13.1
psutil==5.7.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.2.0
PyJWT==1.4.2
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.9
URLObject==2.4.3
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
yapf==0.29.0
